requirements.txt.txt
streamlit
torch
numpy
pandas
transformers
biopython
xgboost
scikit-learn